+++ src/main/java/com/epam/figure/Cylinder.java
@@ -4,17 +4,17 @@ package com.epam.figure;
  * Created by Sergey_Stefoglo on 7/17/2016.
  */
 public class Cylinder {
-    private   int radius;
-    private  int height;
+    private   double radius;
+    private  double height;
 
-    public int getRadius() {
+    public double getRadius() {
         return radius;
     }
-    public int getHeight() {
+    public double getHeight() {
         return height;
     }
 
-    public Cylinder(int radius, int height) {
+    public Cylinder(double radius, double height) {
         this.radius = radius;
         this.height = height;
     }
