+++ src/main/java/com/epam/figure/Parallelepiped.java
@@ -5,19 +5,19 @@ package com.epam.figure;
  */
 public class Parallelepiped {
 
-    private int height;
-    private int squareBase;
+    private double height;
+    private double squareBase;
 
-    public Parallelepiped(int h, int sBase) {
+    public Parallelepiped(double h, double sBase) {
         this.height = h;
         this.squareBase = sBase;
     }
 
-    public int getHeight() {
+    public double getHeight() {
         return height;
     }
 
-    public int getSquareBase() {
+    public double getSquareBase() {
         return squareBase;
     }
 }
