+++ src/main/java/com/epam/figure/Tetrahedron.java
@@ -4,13 +4,13 @@ package com.epam.figure;
  * Created by Sergey_Stefoglo on 7/17/2016.
  */
 public class Tetrahedron {
-    private int edgeA;
+    private double edgeA;
 
-    public int getEdgeA() {
+    public double getEdgeA() {
         return edgeA;
     }
 
-    public Tetrahedron(int a) {
+    public Tetrahedron(double a) {
         this.edgeA = a;
     }
 }
